+++ src/Support/helpers.php
@@ -11,6 +11,7 @@ declare(strict_types=1);
 
 use Konekt\AppShell\Contracts\Widget;
 use Konekt\AppShell\EnumColors;
+use Konekt\AppShell\Helpers\ColorHelper;
 use Konekt\AppShell\Theme\ThemeColor;
 use Konekt\AppShell\Widgets;
 use Konekt\Enum\Enum;
@@ -49,6 +50,13 @@ function avatar_image_url($model = null, int $size = 100)
 function enum_color(Enum $enum): ThemeColor|string
 {
     return EnumColors::colorOf($enum);
+}
+
+if (!function_exists('coloring')) {
+    function coloring(): ColorHelper
+    {
+        return helper('color');
+    }
 }
 
 if (!function_exists('show_date')) {