+++ src/resources/views/components/badge.blade.php
@@ -1,0 +1,1 @@
+<span @class(["badge text-bg-{$variant}", "fs-$fontSize" => $fontSize])>{{ $slot }}</span>