+++ src/resources/views/components/card.blade.php
@@ -1,6 +1,6 @@
 <div {{ $attributes->class(['card', "card-accent-$accent" => $accent, 'mb-4']) }}>
     @if(isset($title) || isset($actions))
-        <div class="card-header my-1    ">
+        <div class="card-header my-1">
             <h5 class="card-title pb-1" {{ $title->attributes }}>{{ $title }}</h5>
             @isset($actions)
                 <div class="card-actionbar" {!! $actions->attributes !!}>