+++ src/resources/views/user/create.blade.php
@@ -9,23 +9,21 @@
 {!! Form::model($user, ['route' => 'appshell.user.store', 'autocomplete' => 'off', 'class' => 'row mb-3']) !!}
 
 <div class="col-12 col-md-6 col-lg-8 col-xl-9">
-    @component(theme_widget('group'), ['accent' => 'success'])
-        @slot('title'){{ __('Enter Account Details') }}@endslot
-
+    <x-appshell::card accent="success">
+        <x-slot:title>{{ __('Enter Account Details') }}</x-slot:title>
         @include('appshell::user._form')
-
-        @slot('footer')
-                <button class="btn btn-success">{{ __('Create user') }}</button>
-                <a href="#" onclick="history.back();" class="btn btn-link text-muted">{{ __('Cancel') }}</a>
-        @endslot
-    @endcomponent
+        <x-slot:footer>
+            <x-appshell::button variant="success">{{ __('Create user') }}</x-appshell::button>
+            <x-appshell::button type="button" onclick="history.back();" variant="link" class="text-muted">{{ __('Cancel') }}</x-appshell::button>
+        </x-slot:footer>
+    </x-appshell::card>
 </div>
 
 <div class="col-12 col-md-6 col-lg-4 col-xl-3">
-    @component(theme_widget('group'), ['accent' => 'success'])
-        @slot('title'){{ __('Roles') }}@endslot
+    <x-appshell::card accent="success">
+        <x-slot:title>{{ __('Roles') }}</x-slot:title>
         @include('appshell::role._assignment', ['model' => $user])
-    @endcomponent
+    </x-appshell::card>
 </div>
 
 {!! Form::close() !!}